backend/services/biometric-service/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3.0
redis>=4.5
numpy>=1.24.0
Pillow>=10.0.0
django-cors-headers>=4.0
